Changed SCR to cut off at a higher level
We are thinking that LED light bulbs are still coming perhaps intermittently when the A/C level is very low.   Since the A/C level is not only low but also at a reduced duty cycle, perhaps the LEDs in the bulb are not acting in a consistent manner.   So line 162 in the interrupt service routine used to be set at 10.   It is now set at 40 which represents around 15 percent of the sine wave.   It is hoped that this new level will correct any low-level inconsistencies of the LED light bulbs.